genai-llm-rag-bedrock-poc/requirements.txt
streamlit
Pillow
PyPDF2>=3.0.0
pytesseract
boto3
langchain-aws
langdetect
